Add ReducedTimelineSchedule
